langgraph_workflows/dhg-agents-cloud/src/dhg-audio-agent/requirements.txt
# DHG Audio Analysis Agent — Python Dependencies
# Per Build Spec Section 2.2

# Core ML
faster-whisper>=1.0.0
pyannote.audio>=3.1
torch>=2.0

# LangGraph
langgraph>=0.2
langchain-core>=0.3

# API
fastapi>=0.110
uvicorn>=0.29
httpx>=0.27
pydantic>=2.6
pydantic-settings>=2.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0
asyncpg>=0.29
pgvector>=0.3

# Audio processing
librosa>=0.10
soundfile>=0.12

# Observability
prometheus-client>=0.20

# Utilities
python-dotenv>=1.0
